chore: refactor code, remove unless code (#1789)
* refactor: remove code for build&start

* refactor: merge demo  generatelogic

* refactor: remove unused files

* refactor: generate

* fix: exit when generate  error

* fix: denpencies

* revert: generate

* revert: generate

* fix: lint
