+++ tools/ice-devtools/package.json
@@ -13,31 +13,9 @@
   },
   "license": "MIT",
   "dependencies": {
-    "@alifd/dts-generator": "^1.0.3",
-    "@babel/core": "7.1.2",
-    "@babel/plugin-proposal-class-properties": "7.1.0",
-    "@babel/plugin-proposal-decorators": "7.1.2",
-    "@babel/plugin-proposal-do-expressions": "7.0.0",
-    "@babel/plugin-proposal-export-default-from": "7.0.0",
-    "@babel/plugin-proposal-export-namespace-from": "7.0.0",
-    "@babel/plugin-proposal-function-bind": "7.0.0",
-    "@babel/plugin-proposal-function-sent": "7.1.0",
-    "@babel/plugin-proposal-json-strings": "7.0.0",
-    "@babel/plugin-proposal-logical-assignment-operators": "7.0.0",
-    "@babel/plugin-proposal-nullish-coalescing-operator": "7.0.0",
-    "@babel/plugin-proposal-numeric-separator": "7.0.0",
-    "@babel/plugin-proposal-optional-chaining": "7.0.0",
-    "@babel/plugin-proposal-pipeline-operator": "7.0.0",
-    "@babel/plugin-proposal-throw-expressions": "7.0.0",
-    "@babel/plugin-syntax-dynamic-import": "7.0.0",
-    "@babel/plugin-syntax-import-meta": "7.0.0",
-    "@babel/preset-env": "7.1.0",
-    "@babel/preset-react": "7.0.0",
+    "@babel/core": "^7.4.3",
     "async": "^2.6.1",
-    "awesome-typescript-loader": "^5.2.1",
     "axios": "^0.18.0",
-    "babel-loader": "^8.0.5",
-    "babel-plugin-import": "^1.11.0",
     "babel-plugin-transform-es2015-modules-commonjs": "^6.26.2",
     "babel-plugin-transform-export-extensions": "^6.22.0",
     "babel-plugin-transform-lib-import": "^0.1.0",
@@ -47,10 +25,9 @@
     "commander": "^2.19.0",
     "consolidate": "^0.15.1",
     "cross-spawn": "^6.0.5",
-    "css-loader": "^1.0.1",
     "debug": "^2.6.3",
+    "decamelize": "^3.2.0",
     "easyfile": "^0.2.0",
-    "file-loader": "^3.0.1",
     "handlebars": "^4.0.12",
     "handlebars-loader": "^1.7.0",
     "html-webpack-plugin": "^3.2.0",
@@ -64,43 +41,26 @@
     "less-loader": "^4.1.0",
     "marked": "^0.6.0",
     "metalsmith": "^2.3.0",
-    "mini-css-extract-plugin": "^0.5.0",
     "minimatch": "^3.0.4",
     "mkdirp": "^0.5.1",
     "moment": "^2.22.2",
     "multimatch": "^2.1.0",
-    "node-sass": "^4.11.0",
     "npm-updater": "^3.0.3",
-    "optimize-css-assets-webpack-plugin": "^5.0.1",
     "ora": "^3.0.0",
     "prismjs": "^1.15.0",
-    "progress-bar-webpack-plugin": "^1.11.0",
     "props-schema-generator": "^0.1.1",
     "puppeteer": "^1.11.0",
     "read-metadata": "^1.0.0",
     "request": "^2.88.0",
     "request-promise-native": "^1.0.5",
     "rimraf": "^2.6.3",
-    "sass-loader": "^7.1.0",
     "semver": "^5.6.0",
     "tar": "^2.2.1",
     "tildify": "^1.2.0",
     "typescript": "^3.3.3333",
-    "uglifyjs-webpack-plugin": "^2.1.2",
     "uppercamelcase": "^3.0.0",
-    "url-loader": "^1.1.2",
     "user-home": "^2.0.0",
     "validate-npm-package-name": "^3.0.0",
-    "vue-loader": "^15.5.1",
-    "vue-style-loader": "^4.1.2",
-    "webpack": "^4.27.0",
-    "webpack-bundle-analyzer": "^3.0.4",
-    "webpack-chain": "^5.2.0",
-    "webpack-dev-server": "^3.1.10",
-    "webpack-format-messages": "^2.0.5",
-    "webpack-hot-client": "^4.1.1",
-    "webpack-plugin-import": "^0.2.1",
-    "webpack-simple-progress-plugin": "^0.0.4",
     "zlib": "^1.0.5"
   },
   "engines": {
