--- tools/ice-devtools/template/component/index.js.hbs
@@ -1,3 +1,0 @@
-{{#each demos}}
-  require('{{path}}');
-{{/each}}
