--- tools/ice-devtools/template/preview/block-react-index.js.hbs
@@ -1,9 +1,0 @@
-import React from '{{blockNodeModulesPath}}react';
-import ReactDOM from '{{blockNodeModulesPath}}react-dom'
-import Block from '{{entry}}';
-
-const mountNode = document.querySelector('#mountNode');
-ReactDOM.render(
-  <Block />,
-  mountNode
-);
