--- tools/ice-devtools/utils/compile-es5.js
@@ -1,15 +1,0 @@
-/* Deprecated */
-/* eslint-disable */
-const babel = require('@babel/core');
-const getBabelConfig = require('../config/getBabelConfig');
-
-module.exports = function (source) {
-  const babelrc = getBabelRC();
-  const { code } = babel.transform(source, babelrc);
-  return code;
-};
-
-function getBabelRC() {
-  const babelConfig = getBabelConfig();
-  return babelConfig;
-}
