--- tools/ice-devtools/utils/format-path-for-win.js
@@ -1,4 +1,0 @@
-module.exports = (path) => {
-  const isWin = process.platform === 'win32';
-  return isWin ? path.replace(/\\/g, '\\\\') : path;
-};
