--- tools/ice-devtools/utils/internal-library.js
@@ -1,10 +1,0 @@
-// 这些仓库是需要做自动的样式引入的
-module.exports = [
-  /@icedesign\/.*/,
-  /^@icedesign\/base\/lib\/([^/]+)/,
-  /@alife\/.*/,
-  /^@alife\/next\/lib\/([^/]+)/,
-  /@alifd\/.*/,
-  /^@alifd\/next\/lib\/([^/]+)/,
-  /@ali\/ice-.*/,
-];
